更新：增加 --dry-run 和日志相关选项

+++ source/PotatoPeeler/src/main/java/indi/somebottle/tasks/RegionTaskRunner.java
@@ -59,21 +59,24 @@ public class RegionTaskRunner implements Runnable {
                 region = RegionUtils.readRegion(mcaFile);
             } catch (Exception e) {
                 // 读取失败时检查有没有 .mca.bak 文件，如果有就尝试读取 .mca.bak
-                GlobalLogger.severe("Exception occurred while reading region file: " + mcaFile.getAbsolutePath(), e);
+                GlobalLogger.warning("Exception occurred while reading region file: " + mcaFile.getAbsolutePath(), e);
                 // 检查有没有 .mca.bak 文件
                 if (backupFile.exists()) {
                     GlobalLogger.info("Backup file found. Trying to read backup file: " + backupFile.getAbsolutePath());
                     try {
                         // 如果有的话尝试读取 .mca.bak
                         region = RegionUtils.readRegion(backupFile);
-                        // 把无法读取的 .mca 移除，然后把备份文件重命名为 .mca，方便进行后面的流程
-                        if (mcaFile.exists() && !mcaFile.delete()) {
-                            GlobalLogger.warning("Failed to delete file: " + mcaFile.getAbsolutePath());
-                            continue;
+                        // dryRun 模式下不执行实际的写操作
+                        if (!params.dryRun) {
+                            // 把无法读取的 .mca 移除，然后把备份文件重命名为 .mca，方便进行后面的流程
+                            if (mcaFile.exists() && !mcaFile.delete()) {
+                                GlobalLogger.warning("Failed to delete file: " + mcaFile.getAbsolutePath());
+                                continue;
+                            }
+                            Files.move(backupMCAPath, originalMCAPath);
                         }
-                        Files.move(backupMCAPath, originalMCAPath);
                     } catch (Exception ex) {
-                        GlobalLogger.severe("Exception occurred while reading and restoring backup region file: " + backupFile.getAbsolutePath(), e);
+                        GlobalLogger.warning("Exception occurred while reading backup region file: " + backupFile.getAbsolutePath(), e);
                         continue;
                     }
                 } else {
@@ -107,45 +110,57 @@ public class RegionTaskRunner implements Runnable {
             if (!hasModified) {
                 continue;
             }
-            // 先把原文件重命名为 .mca.bak 后缀
-            try {
-                // 先检查目标 backup 文件是否存在，若存在则移除
-                if (backupFile.exists() && !backupFile.delete()) {
-                    GlobalLogger.warning("Failed to delete previous backup file: " + backupFile.getAbsolutePath());
+            if (params.dryRun) {
+                // ------------- Dry-run -------------
+                try {
+                    // 模拟写入区域文件
+                    long bytesWrite = RegionUtils.writeRegion(region, mcaFile, null, true);
+                    sizeReduced += (originalLength - bytesWrite);
+                } catch (IOException e) {
+                    GlobalLogger.warning("Failed to write modified region to file(dry-run): ", e);
                     continue;
                 }
-                // 把原文件临时重命名为 .mca.bak
-                Files.move(originalMCAPath, backupMCAPath);
-            } catch (IOException e) {
-                // 文件重命名失败
-                GlobalLogger.severe("Failed to backup(rename) file:" + mcaFile.getAbsolutePath(), e);
-                continue;
-            }
-            // 把修改后的区域写回原 mcaFile
-            try {
-                // 注意，到这里时原 .mca 文件已经被重命名了 .mca.bak
-                // 因此要从 .mca.bak 拷贝数据到新生成的 .mca
-                RegionUtils.writeRegion(region, backupFile, mcaFile);
-            } catch (IOException e) {
-                // 写入失败，恢复备份
-                GlobalLogger.severe("Failed to write modified region to file: " + mcaFile.getAbsolutePath() + ", restoring backup...", e);
+            } else {
+                // ------------- 实际运行 -------------
                 try {
-                    // 如果 mcaFile 存在则尝试删除
-                    if (mcaFile.exists() && !mcaFile.delete()) {
-                        GlobalLogger.severe("Failed to delete file: " + mcaFile.getAbsolutePath(), e);
+                    // 先检查目标 backup 文件是否存在，若存在则移除
+                    if (backupFile.exists() && !backupFile.delete()) {
+                        GlobalLogger.warning("Failed to delete previous backup file: " + backupFile.getAbsolutePath());
+                        continue;
                     }
-                    Files.move(backupMCAPath, originalMCAPath);
-                } catch (IOException ex) {
-                    GlobalLogger.severe("Unexpected! Failed to restore backup file: " + backupMCAPath, ex);
+                    // 把原文件临时重命名为 .mca.bak
+                    Files.move(originalMCAPath, backupMCAPath);
+                } catch (IOException e) {
+                    // 文件重命名失败
+                    GlobalLogger.warning("Failed to backup(rename) file:" + mcaFile.getAbsolutePath(), e);
+                    continue;
                 }
-                continue;
+                // 把修改后的区域写回原 mcaFile
+                try {
+                    // 注意，到这里时原 .mca 文件已经被重命名了 .mca.bak
+                    // 因此要从 .mca.bak 拷贝数据到新生成的 .mca
+                    RegionUtils.writeRegion(region, backupFile, mcaFile, false);
+                } catch (IOException e) {
+                    // 写入失败，恢复备份
+                    GlobalLogger.warning("Failed to write modified region to file: " + mcaFile.getAbsolutePath() + ", restoring backup...", e);
+                    try {
+                        // 如果 mcaFile 存在则尝试删除
+                        if (mcaFile.exists() && !mcaFile.delete()) {
+                            GlobalLogger.severe("Failed to delete file: " + mcaFile.getAbsolutePath(), e);
+                        }
+                        Files.move(backupMCAPath, originalMCAPath);
+                    } catch (IOException ex) {
+                        GlobalLogger.severe("Unexpected! Failed to restore backup file: " + backupMCAPath, ex);
+                    }
+                    continue;
+                }
+                // 若成功写入则删除备份
+                if (!backupFile.delete()) {
+                    GlobalLogger.warning("Failed to delete backup file: " + backupFile.getAbsolutePath());
+                }
+                // 统计减少的数据大小
+                sizeReduced += (originalLength - mcaFile.length());
             }
-            // 若成功写入则删除备份
-            if (!backupFile.delete()) {
-                GlobalLogger.warning("Failed to delete backup file: " + backupFile.getAbsolutePath());
-            }
-            // 统计
-            sizeReduced += (originalLength - mcaFile.length());
             regionsAffected++;
         }
         // 更新任务结果
